+++ src/commonTest/kotlin/kotools/types/experimental/NotBlankStringTest.kt
@@ -10,8 +10,8 @@ import kotools.types.text.toNotBlankString
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
+@ExperimentalKotoolsTypesApi
 class NotBlankStringTest {
-    @ExperimentalKotoolsTypesApi
     @Test
     fun plus_should_pass_with_a_String() {
         val first: NotBlankString = "hello".toNotBlankString()
@@ -23,13 +23,23 @@ class NotBlankStringTest {
         assertEquals(expected, actual)
     }
 
-    @ExperimentalKotoolsTypesApi
     @Test
     fun plus_should_pass_with_a_NotBlankString() {
         val first: NotBlankString = "hello".toNotBlankString()
             .getOrThrow()
         val second: NotBlankString = " world".toNotBlankString()
             .getOrThrow()
+        val actual: NotBlankString = first + second
+        val expected: NotBlankString = "$first$second".toNotBlankString()
+            .getOrThrow()
+        assertEquals(expected, actual)
+    }
+
+    @Test
+    fun plus_should_pass_with_a_Char() {
+        val first: NotBlankString = "hell".toNotBlankString()
+            .getOrThrow()
+        val second = 'o'
         val actual: NotBlankString = first + second
         val expected: NotBlankString = "$first$second".toNotBlankString()
             .getOrThrow()